detecting escrow closed event in ERC20FeeProxy handlers
