+++ scripts/prepare.ts
@@ -46,6 +46,7 @@ type DataSource = {
     eventSignature: string;
     handlerName: string;
   }[];
+  graphEntities: string[];
 };
 
 const getArtifactInfo = (artifact: ContractArtifact<any>, network: string) => {
@@ -75,7 +76,16 @@ for (const [pn, artifact] of Object.entries(paymentNetworks)) {
 
 for (const network of networks) {
   const dataSources: DataSource[] = [];
+  console.log(`parsing network ${network}`);
+
   Object.entries(paymentNetworks).forEach(([pn, artifact]) => {
+    let graphEntities: string[];
+    if (pn === "ERC20EscrowToPay") {
+      graphEntities = ["Payment", "Escrow", "EscrowEvent"];
+    } else {
+      graphEntities = ["Payment"];
+    }
+
     const infoArray = getArtifactInfo(artifact, network);
     infoArray.forEach(({ address, creationBlockNumber, version }, i) => {
       const events = artifact
@@ -97,6 +107,7 @@ for (const network of networks) {
         address,
         creationBlockNumber,
         events,
+        graphEntities,
       });
     });
   });
