+++ src/shared.ts
@@ -1,4 +1,5 @@
 import { ByteArray, Bytes, crypto, ethereum } from "@graphprotocol/graph-ts";
+import { EscrowEvent } from "../generated/schema";
 
 export function generateId(
   transaction: ethereum.Transaction,
@@ -10,4 +11,23 @@ export function generateId(
 
 export function generateEscrowId(paymentReference: Bytes): string {
   return paymentReference.toHex();
+}
+
+export function createEscrowEvent(
+  event: ethereum.Event,
+  paymentReference: Bytes
+): EscrowEvent {
+  let escrowEvent = new EscrowEvent(
+    generateId(event.transaction, paymentReference)
+  );
+  escrowEvent.reference = paymentReference;
+  escrowEvent.contractAddress = event.address;
+  escrowEvent.from = event.transaction.from;
+  escrowEvent.block = event.block.number.toI32();
+  escrowEvent.timestamp = event.block.timestamp.toI32();
+  escrowEvent.txHash = event.transaction.hash;
+  escrowEvent.gasUsed = event.transaction.gasUsed;
+  escrowEvent.gasPrice = event.transaction.gasPrice;
+  escrowEvent.escrow = generateEscrowId(paymentReference);
+  return escrowEvent;
 }
